#### What this PR does
In https://github.com/grafana/mimir/pull/5665 we reduced active series tracker stripes from 512 to 128. Benchmarks didn't show any difference, but I've noticed we don't have a concurrency benchmark leveraging on the stripes, because the existing concurrency benchmark always use the same series, so always touching the same stripe.

In this PR I'm introducing a concurrency test using multiple series. I've compared 512 (before) vs 128 (after) stripes and there's no practical difference. I've run the benchmarks multiple times and always go some variance, sometimes a bit faster, sometimes a bit slower, but IMO due to "disturbance" and not by a real change in performance:

```
goos: darwin
goarch: amd64
pkg: github.com/grafana/mimir/pkg/ingester/activeseries
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
                                                                            │  before.txt  │              after.txt              │
                                                                            │    sec/op    │    sec/op     vs base               │
ActiveSeries_UpdateSeriesConcurrency/Single_series,_concurrency_=_50-12       53.45n ± ∞ ¹   51.85n ± ∞ ¹        ~ (p=0.548 n=5)
ActiveSeries_UpdateSeriesConcurrency/Single_series,_concurrency_=_100-12      51.21n ± ∞ ¹   55.43n ± ∞ ¹   +8.24% (p=0.032 n=5)
ActiveSeries_UpdateSeriesConcurrency/Single_series,_concurrency_=_500-12      48.88n ± ∞ ¹   58.06n ± ∞ ¹  +18.78% (p=0.008 n=5)
ActiveSeries_UpdateSeriesConcurrency/Single_series,_concurrency_=_1000-12     49.38n ± ∞ ¹   54.32n ± ∞ ¹  +10.00% (p=0.008 n=5)
ActiveSeries_UpdateSeriesConcurrency/Multiple_series,_concurrency_=_50-12     35.07n ± ∞ ¹   38.13n ± ∞ ¹        ~ (p=0.151 n=5)
ActiveSeries_UpdateSeriesConcurrency/Multiple_series,_concurrency_=_100-12    34.65n ± ∞ ¹   35.78n ± ∞ ¹        ~ (p=0.421 n=5)
ActiveSeries_UpdateSeriesConcurrency/Multiple_series,_concurrency_=_500-12    35.17n ± ∞ ¹   30.10n ± ∞ ¹  -14.42% (p=0.008 n=5)
ActiveSeries_UpdateSeriesConcurrency/Multiple_series,_concurrency_=_1000-12   35.52n ± ∞ ¹   29.09n ± ∞ ¹  -18.10% (p=0.008 n=5)
geomean                                                                       42.19n         42.59n         +0.96%
¹ need >= 6 samples for confidence interval at level 0.95

                                                                            │ before.txt  │              after.txt              │
                                                                            │    B/op     │    B/op      vs base                │
ActiveSeries_UpdateSeriesConcurrency/Single_series,_concurrency_=_50-12       0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
ActiveSeries_UpdateSeriesConcurrency/Single_series,_concurrency_=_100-12      0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
ActiveSeries_UpdateSeriesConcurrency/Single_series,_concurrency_=_500-12      0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
ActiveSeries_UpdateSeriesConcurrency/Single_series,_concurrency_=_1000-12     0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
ActiveSeries_UpdateSeriesConcurrency/Multiple_series,_concurrency_=_50-12     0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
ActiveSeries_UpdateSeriesConcurrency/Multiple_series,_concurrency_=_100-12    0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
ActiveSeries_UpdateSeriesConcurrency/Multiple_series,_concurrency_=_500-12    0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
ActiveSeries_UpdateSeriesConcurrency/Multiple_series,_concurrency_=_1000-12   0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
geomean                                                                                 ³                +0.00%               ³
¹ need >= 6 samples for confidence interval at level 0.95
² all samples are equal
³ summaries must be >0 to compute geomean

                                                                            │ before.txt  │              after.txt              │
                                                                            │  allocs/op  │  allocs/op   vs base                │
ActiveSeries_UpdateSeriesConcurrency/Single_series,_concurrency_=_50-12       0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
ActiveSeries_UpdateSeriesConcurrency/Single_series,_concurrency_=_100-12      0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
ActiveSeries_UpdateSeriesConcurrency/Single_series,_concurrency_=_500-12      0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
ActiveSeries_UpdateSeriesConcurrency/Single_series,_concurrency_=_1000-12     0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
ActiveSeries_UpdateSeriesConcurrency/Multiple_series,_concurrency_=_50-12     0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
ActiveSeries_UpdateSeriesConcurrency/Multiple_series,_concurrency_=_100-12    0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
ActiveSeries_UpdateSeriesConcurrency/Multiple_series,_concurrency_=_500-12    0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
ActiveSeries_UpdateSeriesConcurrency/Multiple_series,_concurrency_=_1000-12   0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
geomean                                                                                 ³                +0.00%               ³
¹ need >= 6 samples for confidence interval at level 0.95
² all samples are equal
³ summaries must be >0 to compute geomean
```

#### Which issue(s) this PR fixes or relates to

Fixes #<issue number>

#### Checklist

- [ ] Tests updated
- [ ] Documentation added
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`